# api/openapi.yaml
openapi: 3.0.3
info:
  title: BSV21 Overlay API
  description: |
    API for BSV21 token transactions, providing real-time event streaming, lookups, and SPV validation.
    
    This service combines BSV21-specific operations with core overlay service functionality.
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/b-open-io/bsv21-overlay
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
paths:
...
  /api/1sat/events/{event}:
    get:
      tags:
        - 1sat-events
      summary: Query events by type
      description: Returns all outputs matching the specified event type
      parameters:
        - name: event
          in: path
          required: true
          schema:
            type: string
          description: Event type to query (e.g., "bsv21:tokenId", "p2pkh:address")
        - name: from
          in: query
          schema:
            type: number
            format: float
          description: Starting score for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Maximum number of results to return
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BSV21EventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/1sat/events/{event}/unspent:
    get:
      tags:
        - 1sat-events
      summary: Query unspent events by type
      description: Returns only unspent outputs matching the specified event type
      parameters:
        - name: event
          in: path
          required: true
          schema:
            type: string
          description: Event type to query
        - name: from
          in: query
          schema:
            type: number
            format: float
          description: Starting score for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Maximum number of results to return
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BSV21EventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/1sat/bsv21/{tokenId}:
    get:
      tags:
        - 1sat-bsv21
      summary: Get token events
      description: Returns all events for a specific BSV21 token
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
          description: Token ID to query
        - name: from
          in: query
          schema:
            type: number
            format: float
          description: Starting score for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Maximum number of results to return
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BSV21EventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/1sat/bsv21/{tokenId}/block/{height}:
    get:
      tags:
        - 1sat-bsv21
      summary: Get token transactions at block height
      description: Returns all BSV21 token transactions at a specific block height
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
          description: Token ID to query
        - name: height
          in: path
          required: true
          schema:
            type: integer
          description: Block height to query
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionData'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/1sat/bsv21/{tokenId}/{lockType}/{address}/balance:
    get:
      tags:
        - 1sat-bsv21
      summary: Get token balance for address
      description: Returns the BSV21 token balance for a specific address and lock type
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
          description: Token ID
        - name: lockType
          in: path
          required: true
          schema:
            type: string
            enum: [p2pkh, ltm, pow20]
          description: Lock type
        - name: address
          in: path
          required: true
          schema:
            type: string
          description: Address to query balance for
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BSV21BalanceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/1sat/bsv21/{tokenId}/{lockType}/balance:
    post:
      tags:
        - 1sat-bsv21
      summary: Get token balances for multiple addresses
      description: Returns BSV21 token balances for multiple addresses of a specific lock type
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
          description: Token ID
        - name: lockType
          in: path
          required: true
          schema:
            type: string
            enum: [p2pkh, ltm, pow20]
          description: Lock type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                addresses:
                  type: array
                  items:
                    type: string
                  description: List of addresses to query
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/BSV21BalanceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/1sat/block/tip:
    get:
      tags:
        - 1sat-blocks
      summary: Get current block tip
      description: Returns the current blockchain tip information
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockHeader'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/1sat/block/{height}:
    get:
      tags:
        - 1sat-blocks
      summary: Get block by height
      description: Returns block information for a specific height
      parameters:
        - name: height
          in: path
          required: true
          schema:
            type: integer
          description: Block height
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockHeader'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/1sat/subscribe/{topics}:
    get:
      tags:
        - 1sat-events
      summary: Subscribe to real-time events
      description: Server-sent events stream for real-time BSV21 events
      parameters:
        - name: topics
          in: path
          required: true
          schema:
            type: string
          description: Comma-separated list of topics to subscribe to
      responses:
        '200':
          description: Server-sent event stream
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE stream of events
        '400':
          $ref: '#/components/responses/BadRequest'

  # Core overlay service routes
...
components:
  schemas:
...
    BSV21EventResponse:
      type: object
      properties:
        outpoint:
          type: string
          example: "ae59f3b898ec61acbdb6cc7a245fabeded0c094bf046f35206a3aec60ef88127.0"
          description: Transaction outpoint (txid.vout)
        score:
          type: number
          format: float
          example: 902578.000001234
          description: Event score for ordering/pagination
        data:
          type: object
          additionalProperties: true
          description: Event-specific data
    BSV21BalanceResponse:
      type: object
      properties:
        balance:
          type: integer
          format: int64
          example: 1000000
          description: Token balance
        utxoCount:
          type: integer
          example: 5
          description: Number of UTXOs
    TransactionData:
      type: object
      properties:
        txid:
          type: string
          description: Transaction ID
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/OutputData'
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/OutputData'
    OutputData:
      type: object
      properties:
        txid:
          type: string
          description: Source transaction ID (only for inputs)
        vout:
          type: integer
          description: Output index
        data:
          type: object
          additionalProperties: true
          description: Protocol-specific data
        script:
          type: string
          format: byte
          description: Locking script
        satoshis:
          type: integer
          format: int64
          description: Output value in satoshis
    BlockHeader:
      type: object
      properties:
        height:
          type: integer
          description: Block height
        hash:
          type: string
          description: Block hash
        prevHash:
          type: string
          description: Previous block hash
        merkleRoot:
          type: string
          description: Merkle root
        timestamp:
          type: integer
          format: int64
          description: Block timestamp
        bits:
          type: integer
          description: Difficulty bits
        nonce:
          type: integer
          description: Nonce

    # Overlay service schemas
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Human-readable error message
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: |
        The server cannot process the request due to malformed syntax or invalid input.
        This error is typically caused by client-side issues, such as missing required parameters,
        invalid data formats, or incorrect request structures.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: |
        The requested resource could not be found. This error occurs when the client
        requests an endpoint or resource that does not exist on the server or cannot be located.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: |
        An unexpected condition was encountered on the server, causing the request to fail.
        This error is typically caused by internal issues such as server misconfigurations,
        database failures, or unhandled exceptions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...